pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.118.0",
    "pydantic>=2.11.9",
    "redis-om>=0.3.5",
    "uvicorn[standard]>=0.37.0",
]
